miner_safety_management_system/python/requirements.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
python-dotenv==1.0.1
joblib==1.4.2
matplotlib==3.9.0
seaborn==0.13.2
scipy==1.13.1
orjson==3.10.3
